Correct louvain formula, solve infinite loops (#5713)
Fixes the formulae used to calculate gain and removal cost in calculation
of one level of the Louvain partition tree. Errors here were causing infinite loops
for some cases in the past, see gh-5175 and gh-5704.

This PR also adds test cases to ensure infinite loops are not entered for these cases.

Co-authored-by: Ross Barnowski <rossbar@berkeley.edu>